requirements.txt
Fabric==1.10.0
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
boto==2.33.0
ecdsa==0.11
gevent==1.0.1
greenlet==0.4.5
itsdangerous==0.24
locustio==0.7.2
msgpack-python==0.4.2
paramiko==1.15.1
pycrypto==2.6.1
requests==2.4.3
wsgiref==0.1.2
